rename: RequestContext 에 더 적절한 이름 부터
- Context 는 요청 사이의 동기화되는 정보라는 느낌이 강함.
- 단순히 요청 정보만 담을 것이므로, Info 라는 네이밍을 사용하도록 변경.
