+++ src/main/java/roomescape/common/log/message/LogMessageProvider.java
@@ -3,17 +3,16 @@ package roomescape.common.log.message;
 
 import java.util.Map;
 import org.springframework.http.ResponseEntity;
-import roomescape.common.log.context.RequestContext;
 
 public interface LogMessageProvider {
 
     String getRequestLog(
-            final RequestContext requestContext,
+            final RequestInfo requestInfo,
             final Map<String, Object> handlerArguments
     );
 
     String getResponseLog(
-            final RequestContext requestContext,
+            final RequestInfo requestInfo,
             final ResponseEntity<?> response
     );
 }
